requirements.txt
numpy==1.24.3
tensorflow==2.14.0
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
keras==2.14.0
mediapipe==0.10.21
cvzone==1.6.1
pyttsx3==2.90 